Use root for building repo and use gitignore for build dir

+++ Makefile
@@ -33,14 +33,28 @@ else
 endif
 
 clean-repo:
+ifneq ($(shell id -u), 0)
+	@echo "Clean needs to run under root user."
+	@exit 1
+else
 	rm -rf build/ *.tar *.metadata.yaml
+endif
 
 build-repo: clean-repo
+ifneq ($(shell id -u), 0)
+	@echo "Build repo needs to run under root user."
+	@exit 1
+else
 	mkdir -p $(ROOT_DIR)/build
-	$(LUET) build --all --tree=$(TREE) --destination $(ROOT_DIR)/build --backend $(BACKEND) --concurrency $(CONCURRENCY) --compression $(COMPRESSION)
+	$(LUET) build --no-spinner --all --tree=$(TREE) --destination $(ROOT_DIR)/build --backend $(BACKEND) --concurrency $(CONCURRENCY) --compression $(COMPRESSION)
+endif
 
 create-repo:
-	$(LUET) create-repo --tree "$(TREE)" \
+ifneq ($(shell id -u), 0)
+	@echo "Create repo need to run under root user."
+	@exit 1
+else
+	$(LUET) create-repo --no-spinner --tree "$(TREE)" \
     --output $(ROOT_DIR)/build \
     --packages $(ROOT_DIR)/build \
     --name "luet-mtree" \
@@ -48,7 +62,7 @@ create-repo:
     --urls "http://localhost:8000" \
     --tree-compression gzip \
     --type http
-
+endif
 
 lint: fmt vet
